fix unsigned compare of len to line_len in sprint_realloc_hexstring
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@10149 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/mib.c
@@ -323,7 +323,7 @@ sprint_realloc_hexstring(u_char ** buf, size_t * buf_len, size_t * out_len,
     if (!line_len)
         line_len=len;
 
-    for (; len > line_len; len -= line_len) {
+    for (; (int)len > line_len; len -= line_len) {
         _sprint_hexstring_line(buf, buf_len, out_len, allow_realloc, cp, line_len);
         cp += line_len;
     }
